Update interface name
In case the old and interface names are different, and the
interface has been successfully updated, update the interface name
used for querying. This was causing the interface not be cleaned up
with hyperstart when the pod was destroyed, although the cleanup is not
necessarily a required step as the VM is shutdown soon after.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>