init: 9p: Set a larger blocksize for 9p mount
Currently we don't set a blocksize for the 9p virtio mount point,
so it defaults (to what I suspect will be 4k, but is hard to
pinpoint in the code).
We used to set a blocksize (msize argument) on the 9p mount before
we switched to hyperstart, and testing with that re-instated shows
an I/O performance gain. Thus, re-instate the blocksize argument
on the 9p mount.

Note, the 131072 is currently experimental - working out if there
is an ideal blocksize is a WIP...

Signed-off-by: Graham Whaley <graham.whaley@intel.com>